requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-filter>=23.5
openpyxl>=3.1
pandas>=2.0
django-htmx>=1.17
